220921 Study 로그인 기능 추가 및 User 권한에 따른 게시물 작성 및 읽기

+++ Make_Board/src/main/java/boardproject/controls/user/UserDeleteController.java
@@ -1,8 +1,12 @@
 package boardproject.controls.user;
 
 import java.util.Map;
+
+import javax.servlet.http.HttpSession;
+
 import boardproject.annotation.Component;
 import boardproject.dao.UserDao;
+import boardproject.vo.User;
 import boardproject.bind.DataBinding;
 import boardproject.controls.Controller;
 
@@ -23,6 +27,13 @@ public class UserDeleteController implements Controller, DataBinding {
 	
 	@Override
 	public String execute(Map<String, Object> model) throws Exception {
+		
+		HttpSession session = (HttpSession) model.get("session");
+		User loginUser = (User) session.getAttribute("loginUser");
+		if (loginUser == null || loginUser.getGrade() != 1) {
+			return "/auth/LogInForm.jsp";
+		}
+		
 		userDao.delete((int)model.get("userNo"));
 		
 		return "redirect:list.do";
